Split transfer client tests into four files
