+++ tests/framework/__init__.py
@@ -1,4 +1,5 @@
 from tests.framework.capturedio_testcase import CapturedIOTestCase
+from tests.framework.transfer_client_testcase import TransferClientTestCase
 from tests.framework.tools import (get_fixture_file_dir,
                                    get_client_data, get_user_data)
 
@@ -12,6 +13,7 @@ from tests.framework.constants import (GO_EP1_ID, GO_EP2_ID,
 
 __all__ = [
     "CapturedIOTestCase",
+    "TransferClientTestCase",
 
     "get_fixture_file_dir",
     "get_client_data",
