+++ tests/unit/test_transfer_client.py
@@ -1,112 +1,16 @@
-import re
-import time
-import unittest
 from random import getrandbits
-from datetime import datetime, timedelta
 
 import globus_sdk
-from tests.framework import (CapturedIOTestCase,
-                             get_client_data, get_user_data,
-                             GO_EP1_ID, GO_EP2_ID, GO_EP1_SERVER_ID,
-                             SDKTESTER1A_NATIVE1_TRANSFER_RT,
-                             SDKTESTER2B_NATIVE1_TRANSFER_RT)
+from tests.framework import (TransferClientTestCase, get_user_data,
+                             GO_EP1_ID, GO_EP2_ID, GO_EP1_SERVER_ID)
 from globus_sdk.exc import TransferAPIError
 from globus_sdk.transfer.paging import PaginatedResource
 
 
-def setUpModule():
+class TransferClientTests(TransferClientTestCase):
     """
-    Cleans out any files in ~/.globus/sharing/ on go#ep1 older than an hour
-    TODO: remove this once deleting shared directories does full cleanup
+    Transfer client tests that don't require time intensive setUp and tearDown.
     """
-    ac = globus_sdk.NativeAppAuthClient(
-        client_id=get_client_data()["native_app_client1"]["id"])
-    authorizer = globus_sdk.RefreshTokenAuthorizer(
-        SDKTESTER1A_NATIVE1_TRANSFER_RT, ac)
-    tc = globus_sdk.TransferClient(authorizer=authorizer)
-
-    path = "~/.globus/sharing/"
-    hour_ago = datetime.utcnow() - timedelta(hours=1)
-    filter_string = "last_modified:," + hour_ago.strftime("%Y-%m-%d %H:%M:%S")
-    try:
-        old_files = tc.operation_ls(GO_EP1_ID, path=path, filter=filter_string)
-    except TransferAPIError:  # if a complete cleanup removed the .globus dir
-        return
-
-    kwargs = {"notify_on_succeeded": False, "notify_on_fail": False}
-    ddata = globus_sdk.DeleteData(tc, GO_EP1_ID, **kwargs)
-
-    for item in old_files:
-        ddata.add_item(path + item["name"])
-
-    if len(ddata["DATA"]):
-        tc.submit_delete(ddata)
-
-
-# class that has setUp and tearDown for all transfer client testing classes
-class BaseTransferClientTests(CapturedIOTestCase):
-
-    __test__ = False  # prevents base class from trying to run tests
-
-    @classmethod
-    def setUpClass(self):
-        """
-        Does an auth flow to create an authorized client for
-        sdktester1a and sdktester2b
-        """
-        ac = globus_sdk.NativeAppAuthClient(
-            client_id=get_client_data()["native_app_client1"]["id"])
-
-        authorizer1 = globus_sdk.RefreshTokenAuthorizer(
-            SDKTESTER1A_NATIVE1_TRANSFER_RT, ac)
-        self.tc = globus_sdk.TransferClient(authorizer=authorizer1)
-
-        authorizer2 = globus_sdk.RefreshTokenAuthorizer(
-            SDKTESTER2B_NATIVE1_TRANSFER_RT, ac)
-        self.tc2 = globus_sdk.TransferClient(authorizer=authorizer2)
-
-    def setUp(self):
-        """
-        Creates a list for tracking cleanup of assets created during testing
-        Sets up a test endpoint
-        """
-        super(BaseTransferClientTests, self).setUp()
-        # list of dicts, each containing a function and a list of args
-        # to pass to that function s.t. calling f(*args) cleans an asset
-        self.asset_cleanup = []
-
-        # test endpoint, uses 128 bits of randomness to prevent collision
-        data = {"display_name": "SDK Test Endpoint-" + str(getrandbits(128)),
-                "description": "Endpoint for testing the SDK"}
-        r = self.tc.create_endpoint(data)
-        self.test_ep_id = r["id"]
-        self.asset_cleanup.append({"function": self.tc.delete_endpoint,
-                                   "args": [r["id"]],
-                                   "name": "test_ep"})  # for ease of removal
-
-    def tearDown(self):
-        """
-        Parses created_assets to destroy all assets created during testing
-        """
-        super(BaseTransferClientTests, self).tearDown()
-        # call the cleanup functions with the arguments they were given
-        for cleanup in self.asset_cleanup:
-            cleanup["function"](*cleanup["args"])
-
-    def deleteHelper(self, ep_id, path):
-        """
-        Helper function for cleanup. Deletes by path and endpoint,
-        """
-        kwargs = {"notify_on_succeeded": False}  # prevent email spam
-        ddata = globus_sdk.DeleteData(self.tc, ep_id,
-                                      label="deleteHelper",
-                                      recursive=True, **kwargs)
-        ddata.add_item(path)
-        self.tc.submit_delete(ddata)
-
-
-# class for transfer client tests that don't require time intensive setup
-class TransferClientTests(BaseTransferClientTests):
 
     __test__ = True  # marks sub-class as having tests
 
@@ -1037,985 +941,3 @@ class TransferClientTests(BaseTransferClientTests):
             self.assertEqual(success["DATA_TYPE"], "successful_transfer")
             self.assertIn("source_path", success)
             self.assertIn("destination_path", success)
-
-
-# class for TransferClientTests that require a unique shared endpoint per test
-# since tearDown takes significantly longer if a shared endpoint is deleted
-class SharedTransferClientTests(BaseTransferClientTests):
-
-    __test__ = True  # marks sub-class as having tests
-
-    def setUp(self):
-        """
-        calls BaseTransferClientTest setUp,
-        then sets up a test shared endpoint
-        """
-        super(SharedTransferClientTests, self).setUp()
-
-        # shared endpoint hosted on go#ep1,
-        # name and dir randomized to prevent collision
-        share_path = "/~/share-" + str(getrandbits(128))
-        self.tc.operation_mkdir(GO_EP1_ID, path=share_path)
-        shared_data = {"DATA_TYPE": "shared_endpoint",
-                       "host_endpoint": GO_EP1_ID,
-                       "host_path": share_path,
-                       "display_name":
-                       "SDK Test Shared Endpoint-" + str(getrandbits(128)),
-                       "description": "Shared Endpoint for testing the SDK"
-                       }
-        r = self.tc.create_shared_endpoint(shared_data)
-        self.test_share_ep_id = r["id"]
-
-        # track assets for cleanup
-        self.asset_cleanup.append({"function": self.tc.delete_endpoint,
-                                   "args": [r["id"]]})
-        self.asset_cleanup.append({"function": self.deleteHelper,
-                                   "args": [GO_EP1_ID, share_path]})
-
-    def test_endpoint_search(self):
-        """
-        Searches by fulltext, owner_id, and scopes, validates results
-        """
-
-        # full-text and cap of num_results
-        search_string = "tutorial"
-        cap = 5
-        text_doc = self.tc.endpoint_search(search_string, num_results=cap)
-        # confirm the results are in the right format, are capped correctly,
-        # and that the results have the search string in their display names
-        self.assertIsInstance(text_doc, PaginatedResource)
-        results_count = 0
-        for ep in text_doc:
-            self.assertIn(search_string, str.lower(str(ep["display_name"])))
-            results_count += 1
-        self.assertEqual(results_count, cap)
-
-        # owner-id param
-        params = {"filter_owner_id": get_user_data()["go"]["id"]}
-        owner_doc = self.tc.endpoint_search(**params)
-        # confirm format, and that all results are from GO
-        self.assertIsInstance(owner_doc, PaginatedResource)
-        for ep in owner_doc:
-            self.assertEqual(ep["owner_id"], get_user_data()["go"]["id"])
-
-        # scope: my endpoints
-        my_doc = self.tc.endpoint_search(filter_scope="my-endpoints")
-        # confirm format, and that all results are owned by SDK tester
-        self.assertIsInstance(my_doc, PaginatedResource)
-        for ep in my_doc:
-            self.assertEqual(ep["owner_id"],
-                             get_user_data()["sdktester1a"]["id"])
-
-        # scope: shared endpoints
-        my_doc = self.tc.endpoint_search(filter_scope="shared-by-me")
-        # confirm format, and that all results are shared by SDK tester
-        self.assertIsInstance(my_doc, PaginatedResource)
-        for ep in my_doc:
-            self.assertEqual(ep["owner_id"],
-                             get_user_data()["sdktester1a"]["id"])
-            self.assertIsNotNone(ep["sharing_target_root_path"])
-            self.assertIsNotNone(ep["host_endpoint_id"])
-
-    def test_endpoint_autoactivate(self):
-        """
-        Deactivates, then auto-activates shared endpoint,
-        confirms activation through get,
-        confirms trying again with if_expires_in returns AlreadyActivated
-        """
-        # deactivate
-        self.tc.endpoint_deactivate(self.test_share_ep_id)
-
-        # auto-activate and check for successful response code
-        auto_doc = self.tc.endpoint_autoactivate(self.test_share_ep_id)
-        self.assertEqual(auto_doc["code"],
-                         "AutoActivated.GlobusOnlineCredential")
-
-        # confirm get sees the endpoint as activated now
-        get_doc = self.tc.get_endpoint(self.test_share_ep_id)
-        self.assertTrue(get_doc["activated"])
-
-        # confirm if_expires_in sees the endpoint as activated
-        params = {"if_expires_in": "60"}
-        expires_doc = self.tc.endpoint_autoactivate(
-            self.test_share_ep_id, **params)
-        self.assertEqual(expires_doc["code"], "AlreadyActivated")
-
-        # TODO: test against an endpoint we are not allowed to activate
-
-    def test_endpoint_deactivate(self):
-        """
-        Auto-activates, then deactivates shared endpoint,
-        confirms deactivation through get
-        """
-
-        # activate
-        self.tc.endpoint_autoactivate(self.test_share_ep_id)
-
-        # deactivate and check for successful response code
-        deact_doc = self.tc.endpoint_deactivate(self.test_share_ep_id)
-        self.assertEqual(deact_doc["code"], "Deactivated")
-
-        # confirm get sees the endpoint as deactivated now
-        get_doc = self.tc.get_endpoint(self.test_share_ep_id)
-        self.assertFalse(get_doc["activated"])
-
-    def test_my_shared_endpoint_list(self):
-        """
-        Gets my shared endpoint list, validates results
-        """
-
-        # get shared endpoint list
-        share_doc = self.tc.my_shared_endpoint_list(GO_EP1_ID)
-
-        # confirm doc type, and that test_share_ep_id is on the list
-        self.assertEqual(share_doc["DATA_TYPE"], "endpoint_list")
-        for ep in share_doc:
-            if ep["id"] == self.test_share_ep_id:
-                break
-        else:
-            self.assertFalse("test share not found")
-
-    def test_endpoint_acl_list(self):
-        """
-        Gets endpoint access rule list from test_share_ep, validates results
-        """
-
-        # get endpoint role list
-        list_doc = self.tc.endpoint_acl_list(self.test_share_ep_id)
-
-        # validate responce has been put into
-        self.assertEqual(list_doc["DATA_TYPE"], "access_list")
-        self.assertIn("DATA", list_doc)
-        # confirm that each access rule has some expected values
-        for access in list_doc["DATA"]:
-            self.assertEqual(access["DATA_TYPE"], "access")
-            self.assertIn("id", access)
-            self.assertIn("principal_type", access)
-            self.assertIn("principal", access)
-            self.assertIn("permissions", access)
-
-    def test_get_endpoint_acl_rule(self):
-        """
-        Adds access rule to test_share_ep, gets it by id, validates results
-        """
-
-        # get access rule id through test_add_endpoint_acl_rule
-        access_id = self.test_add_endpoint_acl_rule()
-
-        # get the access rule by id
-        get_doc = self.tc.get_endpoint_acl_rule(self.test_share_ep_id,
-                                                access_id)
-
-        # validate results
-        self.assertEqual(get_doc["DATA_TYPE"], "access")
-        self.assertEqual(get_doc["id"], access_id)
-
-    def test_add_endpoint_acl_rule(self):
-        """
-        Adds access rule to test_share_ep, validates results,
-        then confirms get sees the new access rule,
-        returns the access_id for use in testing get update and delete
-        """
-
-        # add root read access to all users logged in with auth
-        add_data = {"DATA_TYPE": "access",
-                    "principal_type": "all_authenticated_users",
-                    "principal": "",
-                    "path": "/",
-                    "permissions": "r"
-                    }
-        add_doc = self.tc.add_endpoint_acl_rule(self.test_share_ep_id,
-                                                add_data)
-
-        # validate results
-        self.assertEqual(add_doc["DATA_TYPE"], "access_create_result")
-        self.assertEqual(add_doc["code"], "Created")
-        self.assertEqual(add_doc["message"],
-                         "Access rule created successfully.")
-        self.assertIn("access_id", add_doc)
-
-        # confirm get sees new access rule
-        access_id = add_doc["access_id"]
-        get_doc = self.tc.get_endpoint_acl_rule(self.test_share_ep_id,
-                                                access_id)
-        self.assertEqual(get_doc["id"], access_id)
-        for item in add_data:
-            self.assertEqual(get_doc[item], add_data[item])
-
-        # return access_id
-        return access_id
-
-    def test_update_endpoint_acl_rule(self):
-        """
-        Adds access rule to test_share_ep, updates that access rule,
-        validates results, then confirms get sees the new changes
-        """
-
-        # get access rule id through test_add_endpoint_acl_rule
-        access_id = self.test_add_endpoint_acl_rule()
-
-        # update the access rule by id
-        update_data = {"permissions": "rw"}
-        update_doc = self.tc.update_endpoint_acl_rule(self.test_share_ep_id,
-                                                      access_id, update_data)
-
-        # validate results
-        self.assertEqual(update_doc["DATA_TYPE"], "result")
-        self.assertEqual(update_doc["code"], "Updated")
-        self.assertIn("permissions updated successfully",
-                      update_doc["message"])
-
-        # confirm get sees new permissions
-        get_doc = self.tc.get_endpoint_acl_rule(self.test_share_ep_id,
-                                                access_id)
-        self.assertEqual(get_doc["permissions"], update_data["permissions"])
-
-    def test_delete_endpoint_acl_rule(self):
-        """
-        Adds access rule to test_share_ep, deletes that access rule,
-        validates results, then confirms get no longer sees the access rule
-        """
-
-        # get access rule id through test_add_endpoint_acl_rule
-        access_id = self.test_add_endpoint_acl_rule()
-
-        # delete the access rule by id
-        delete_doc = self.tc.delete_endpoint_acl_rule(self.test_share_ep_id,
-                                                      access_id)
-
-        # validate results
-        self.assertEqual(delete_doc["DATA_TYPE"], "result")
-        self.assertEqual(delete_doc["code"], "Deleted")
-        self.assertIn("deleted successfully", delete_doc["message"])
-
-        # confirm get no longer sees the access rule
-        with self.assertRaises(TransferAPIError) as apiErr:
-            self.tc.get_endpoint_acl_rule(self.test_share_ep_id, access_id)
-        self.assertEqual(apiErr.exception.http_status, 404)
-        self.assertEqual(apiErr.exception.code, "AccessRuleNotFound")
-
-    def test_endpoint_manager_create_pause_rule(self):
-        """
-        Creates a pause rule on the shared endpoint, validates results.
-        Returns the rule's id for use in other tests that need a pause rule.
-        Confirms 403 when non manager attempts to use this resource.
-        """
-        rule_data = {
-            "DATA_TYPE": "pause_rule",
-            "message": "SDK Test Pause Rule",
-            "endpoint_id": self.test_share_ep_id,
-            "identity_id": None,  # affect all users on endpoint
-            "start_time": None  # start now
-        }
-        create_doc = self.tc.endpoint_manager_create_pause_rule(rule_data)
-
-        # verify results
-        for key in rule_data:
-            self.assertEqual(create_doc[key], rule_data[key])
-        self.assertTrue(create_doc["editable"])
-        self.assertFalse(create_doc["created_by_host_manager"])
-        rule_id = create_doc["id"]
-
-        # 403 for non managers
-        with self.assertRaises(TransferAPIError) as apiErr:
-            self.tc2.endpoint_manager_create_pause_rule(rule_data)
-        self.assertEqual(apiErr.exception.http_status, 403)
-        self.assertEqual(apiErr.exception.code, "PermissionDenied")
-
-        # track for cleanup, note this must be placed earlier in the
-        # list than the shared endpoint
-        self.asset_cleanup.insert(
-            0,
-            {"function": self.tc.endpoint_manager_delete_pause_rule,
-             "args": [rule_id],
-             "name": "test_pause_rule"})
-
-        # return id for use in other tests
-        return rule_id
-
-    def test_endpoint_manager_pause_rule_list(self):
-        """
-        Gets a pause_rule id from test_endpoint_manager_create_pause_rule.
-        Gets pause rule lists with/without endpoint filters, validates results.
-        Confirms 403 when non manager attempts to use this resource.
-        """
-        rule_id = self.test_endpoint_manager_create_pause_rule()
-
-        # all endpoint_rules
-        list_doc = self.tc.endpoint_manager_pause_rule_list()
-        self.assertEqual(list_doc["DATA_TYPE"], "pause_rule_list")
-        for rule in list_doc:
-            self.assertEqual(rule["DATA_TYPE"], "pause_rule")
-            if rule["id"] == rule_id:
-                break
-        else:
-            self.assertFalse("rule_id not found")
-
-        # filtered by endpoint, should only have the created rule
-        list_doc = self.tc.endpoint_manager_pause_rule_list(
-            filter_endpoint=self.test_share_ep_id)
-        rule = list_doc["DATA"][0]
-        self.assertEqual(rule["id"], rule_id)
-
-        # 403 for non managers
-        with self.assertRaises(TransferAPIError) as apiErr:
-            self.tc2.endpoint_manager_pause_rule_list()
-        self.assertEqual(apiErr.exception.http_status, 403)
-        self.assertEqual(apiErr.exception.code, "PermissionDenied")
-
-    def test_endpoint_manager_get_pause_rule(self):
-        """
-        Gets a pause rule created by test_endpoint_manager_create_pause_rule.
-        Validates results and checks expected fields.
-        Confirms 403 when non manager attempts to use this resource.
-        """
-        rule_id = self.test_endpoint_manager_create_pause_rule()
-        get_doc = self.tc.endpoint_manager_get_pause_rule(rule_id)
-
-        # validate results have expected fields and values
-        expected = {
-            "DATA_TYPE": "pause_rule", "id": rule_id,
-            "message": "SDK Test Pause Rule", "start_time": None,
-            "endpoint_id": self.test_share_ep_id, "identity_id": None,
-            "modified_by_id": get_user_data()["sdktester1a"]["id"],
-            "created_by_host_manager": False, "editable": True,
-            "pause_ls": True, "pause_mkdir": True, "pause_rename": True,
-            "pause_task_delete": True, "pause_task_transfer_write": True,
-            "pause_task_transfer_read": True}
-        for key in expected:
-            self.assertEqual(get_doc[key], expected[key])
-
-        # 403 for non managers
-        with self.assertRaises(TransferAPIError) as apiErr:
-            self.tc2.endpoint_manager_get_pause_rule(rule_id)
-        self.assertEqual(apiErr.exception.http_status, 403)
-        self.assertEqual(apiErr.exception.code, "PermissionDenied")
-
-    def test_endpoint_manager_update_pause_rule(self):
-        """
-        Updates a pause rule created by test_endpoint_manager_create_pause_rule
-        Validates results and confirms fields have updated with get.
-        Confirms 403 when non manager attempts to use this resource.
-        """
-        rule_id = self.test_endpoint_manager_create_pause_rule()
-
-        # update the rule
-        update_data = {
-            "message": "New Message", "pause_ls": False, "pause_mkdir": False,
-            "pause_rename": False, "pause_task_delete": False,
-            "pause_task_transfer_write": False,
-            "pause_task_transfer_read": False}
-        update_doc = self.tc.endpoint_manager_update_pause_rule(
-            rule_id, update_data)
-        self.assertEqual(update_doc["DATA_TYPE"], "pause_rule")
-
-        # get the the rule after update
-        get_doc = self.tc.endpoint_manager_get_pause_rule(rule_id)
-
-        # validate results and confirm get sees updated fields
-        for key in update_data:
-            self.assertEqual(update_doc[key], update_data[key])
-            self.assertEqual(get_doc[key], update_data[key])
-
-        # 403 for non managers
-        with self.assertRaises(TransferAPIError) as apiErr:
-            self.tc2.endpoint_manager_update_pause_rule(rule_id, update_data)
-        self.assertEqual(apiErr.exception.http_status, 403)
-        self.assertEqual(apiErr.exception.code, "PermissionDenied")
-
-    def test_endpoint_manager_delete_pause_rule(self):
-        """
-        Deletes a pause rule created by test_endpoint_manager_create_pause_rule
-        Validates results and confirms 404 when trying to get rule.
-        Confirms 403 when non manager attempts to use this resource.
-        """
-        rule_id = self.test_endpoint_manager_create_pause_rule()
-
-        # confirm 403 for non managers before deleting the rule
-        with self.assertRaises(TransferAPIError) as apiErr:
-            self.tc2.endpoint_manager_delete_pause_rule(rule_id)
-        self.assertEqual(apiErr.exception.http_status, 403)
-        self.assertEqual(apiErr.exception.code, "PermissionDenied")
-
-        # delete the rule, validate results
-        delete_doc = self.tc.endpoint_manager_delete_pause_rule(rule_id)
-        self.assertEqual(delete_doc["DATA_TYPE"], "result")
-        self.assertEqual(delete_doc["code"], "Deleted")
-
-        # confirm get returns 404 as rule no longer exists
-        with self.assertRaises(TransferAPIError) as apiErr:
-            self.tc.endpoint_manager_get_pause_rule(rule_id)
-        self.assertEqual(apiErr.exception.http_status, 404)
-        self.assertEqual(apiErr.exception.code, "PauseRuleNotFound")
-
-        # stop tracking pause rule for cleanup
-        for cleanup in self.asset_cleanup:
-            if "name" in cleanup and cleanup["name"] == "test_pause_rule":
-                self.asset_cleanup.remove(cleanup)
-                break
-
-    def test_task_pause_info(self):
-        """
-        Creates a pause rule on the shared endpoint, then submits a task
-        against it. Gets pause info for the task, validates results,
-        and confirms the task is paused (or about to be).
-        """
-        # create pause rule
-        rule_id = self.test_endpoint_manager_create_pause_rule()
-
-        # submit a no-op delete task
-        ddata = globus_sdk.DeleteData(self.tc, self.test_share_ep_id,
-                                      notify_on_fail=False)
-        ddata.add_item("no-op.txt")
-        task_id = self.tc.submit_delete(ddata)["task_id"]
-
-        # get pause info and validate
-        pause_doc = self.tc.task_pause_info(task_id)
-
-        # validate top level results
-        self.assertEqual(pause_doc["DATA_TYPE"], "pause_info_limited")
-        self.assertIsNone(pause_doc["source_pause_message"])
-        self.assertIsNone(pause_doc["destination_pause_message"])
-
-        # validate the rule results
-        rule = pause_doc["pause_rules"][0]  # should be the only rule
-        self.assertEqual(rule["DATA_TYPE"], "pause_rule_limited")
-        self.assertEqual(rule["id"], rule_id)
-        self.assertEqual(rule["message"], "SDK Test Pause Rule")
-        self.assertNotIn("modified_by", rule)
-        self.assertNotIn("modified_by_id", rule)
-
-    def test_my_effective_pause_rule_list(self):
-        """
-        Creates a pause rule on the shared endpoint, then gets pause rule list.
-        Validates results and confirms the pause rule is found.
-        """
-        # create pause rule
-        rule_id = self.test_endpoint_manager_create_pause_rule()
-
-        # get the pause list
-        pause_doc = self.tc.my_effective_pause_rule_list(self.test_share_ep_id)
-
-        # validate top level results
-        self.assertEqual(pause_doc["DATA_TYPE"], "pause_rule_list")
-
-        # validate the rule results
-        rule = pause_doc["DATA"][0]  # should be the only rule
-        self.assertEqual(rule["DATA_TYPE"], "pause_rule_limited")
-        self.assertEqual(rule["id"], rule_id)
-        self.assertEqual(rule["message"], "SDK Test Pause Rule")
-        self.assertNotIn("modified_by", rule)
-        self.assertNotIn("modified_by_id", rule)
-
-    def test_endpoint_manager_task_pause_info(self):
-        """
-        Creates a pause rule on the shared endpoint, then
-        has sdktester2b submit a no-op task on the shared endpoint.
-        Confirms sdktester1a can see the task is paused (or about to be).
-        Confirms 403 when non manager attempts to use this resource.
-        """
-        # sdktester1a creates pause rule
-        rule_id = self.test_endpoint_manager_create_pause_rule()
-
-        # sdktester2b subits no-op delete task
-        ddata = globus_sdk.DeleteData(self.tc2, self.test_share_ep_id,
-                                      notify_on_fail=False)
-        ddata.add_item("no-op.txt")
-        task_id = self.tc2.submit_delete(ddata)["task_id"]
-
-        # sdktester1a gets the task pause info as admin
-        pause_doc = self.tc.endpoint_manager_task_pause_info(task_id)
-
-        # validate top level results
-        self.assertEqual(pause_doc["DATA_TYPE"], "pause_info_limited")
-        self.assertIsNone(pause_doc["source_pause_message"])
-        self.assertIsNone(pause_doc["destination_pause_message"])
-
-        # validate the rule results
-        rule = pause_doc["pause_rules"][0]  # should be the only rule
-        self.assertEqual(rule["DATA_TYPE"], "pause_rule_limited")
-        self.assertEqual(rule["id"], rule_id)
-        self.assertEqual(rule["message"], "SDK Test Pause Rule")
-        # self.assertEqual(rule["modified_by_id"],
-        #                  get_user_data()["sdktester1a"]["id"])
-
-        # 403 for non managers, even if they submitted the task
-        with self.assertRaises(TransferAPIError) as apiErr:
-            self.tc2.endpoint_manager_task_pause_info(task_id)
-        self.assertEqual(apiErr.exception.http_status, 403)
-        self.assertEqual(apiErr.exception.code, "PermissionDenied")
-
-
-# class for Transfer Client Tests that require the activity_manager
-# effective role on an endpoint but don't require a unique endpoint.
-# Setup checks to see if the managed shared endpoint exists and if not creates
-# one. This endpoint is not removed during cleanup.
-class ManagerTransferClientTests(BaseTransferClientTests):
-
-    __test__ = True  # marks sub-class as having tests
-
-    @classmethod
-    def setUpClass(self):
-        """
-        Sets up a shared endpoint on test gp#ep2 managed by sdktester1a,
-        and shares it with sdktester2b,
-        or sees that this endpoint already exits and gets its id.
-        """
-        super(ManagerTransferClientTests, self).setUpClass()
-
-        try:
-            # shared endpoint hosted on go#ep2 managed by sdktester1a
-            host_path = "/~/managed_ep"
-            self.tc.operation_mkdir(GO_EP2_ID, path=host_path)
-            shared_data = {"DATA_TYPE": "shared_endpoint",
-                           "host_endpoint": GO_EP2_ID,
-                           "host_path": host_path,
-                           "display_name": "SDK Test Managed Endpoint",
-                           "description": "Endpoint for managed SDK testing"
-                           }
-            r = self.tc.create_shared_endpoint(shared_data)
-            self.managed_ep_id = r["id"]
-
-            # share read and write to sdktester2b
-            add_data = {"DATA_TYPE": "access",
-                        "principal_type": "identity",
-                        "principal": get_user_data()["sdktester2b"]["id"],
-                        "path": "/",
-                        "permissions": "rw"}
-            self.tc.add_endpoint_acl_rule(self.managed_ep_id, add_data)
-
-        except TransferAPIError as e:
-            if "already exists" in str(e):
-                shares = self.tc.my_shared_endpoint_list(GO_EP2_ID)
-                self.managed_ep_id = shares["DATA"][0]["id"]
-            else:
-                raise e
-
-    def test_endpoint_manager_monitored_endpoints(self):
-        """
-        Gets a list of all endpoints sdktester1a is an activity_manager on,
-        Confirms list contains managed_ep, and has some expected fields.
-        """
-        ep_doc = self.tc.endpoint_manager_monitored_endpoints()
-        expected_fields = ["my_effective_roles", "display_name",
-                           "owner_id", "owner_string"]
-
-        for ep in ep_doc:
-            self.assertEqual(ep["DATA_TYPE"], "endpoint")
-            for field in expected_fields:
-                self.assertIn(field, ep)
-            if ep["id"] == self.managed_ep_id:
-                break
-        else:
-            self.assertFalse("managed endpoint not found")
-
-    def test_endpoint_manager_get_endpoint(self):
-        """
-        Gets the managed endpoint, confirms expected results
-        Confirms 403 when non manager attempts to use this resource.
-        """
-        ep_doc = self.tc.endpoint_manager_get_endpoint(self.managed_ep_id)
-        self.assertEqual(ep_doc["DATA_TYPE"], "endpoint")
-        self.assertEqual(ep_doc["id"], self.managed_ep_id)
-        self.assertIsNone(ep_doc["in_use"])
-
-        # 403 for non managers
-        with self.assertRaises(TransferAPIError) as apiErr:
-            self.tc2.endpoint_manager_hosted_endpoint_list(self.managed_ep_id)
-        self.assertEqual(apiErr.exception.http_status, 403)
-        self.assertEqual(apiErr.exception.code, "PermissionDenied")
-
-    # TODO: test against a non shared endpoint we have the manager role on
-    def test_endpoint_manager_hosted_endpoint_list(self):
-        """
-        Attempts to gets the list of shares hosted on the managed endpoint.
-        Confirms this fails as shares cannot themselves host shares.
-        Confirms 403 when non manager attempts to use this resource.
-        """
-        with self.assertRaises(TransferAPIError) as apiErr:
-            self.tc.endpoint_manager_hosted_endpoint_list(self.managed_ep_id)
-
-        self.assertEqual(apiErr.exception.http_status, 409)
-        self.assertEqual(apiErr.exception.code, "Conflict")
-        self.assertIn("not a host endpoint", apiErr.exception.message)
-
-        # 403 for non managers
-        with self.assertRaises(TransferAPIError) as apiErr:
-            self.tc2.endpoint_manager_hosted_endpoint_list(self.managed_ep_id)
-        self.assertEqual(apiErr.exception.http_status, 403)
-        self.assertEqual(apiErr.exception.code, "PermissionDenied")
-
-    def test_endpoint_manager_acl_list(self):
-        """
-        Gets ACL list from managed endpoint, validates results
-        Confirms 403 when non manager attempts to use this resource.
-        """
-        list_doc = self.tc.endpoint_manager_acl_list(self.managed_ep_id)
-
-        self.assertEqual(list_doc["DATA_TYPE"], "access_list")
-        expected_fields = ["id", "principal", "principal_type", "permissions"]
-        for access in list_doc["DATA"]:
-            self.assertEqual(access["DATA_TYPE"], "access")
-            for field in expected_fields:
-                self.assertIn(field, access)
-
-        # 403 for non managers
-        with self.assertRaises(TransferAPIError) as apiErr:
-            self.tc2.endpoint_manager_acl_list(self.managed_ep_id)
-        self.assertEqual(apiErr.exception.http_status, 403)
-        self.assertEqual(apiErr.exception.code, "PermissionDenied")
-
-    def test_endpoint_manager_task_list(self):
-        """
-        Has sdktester2b submit transfer and delete task to the managed_ep
-        Then has sdktester1a get its endpoint manager task list
-        Confirms tasks submitted by sdktester2b on the managed endpoint
-        are visible, and some expected fields are present.
-        """
-        # sdktester2b submits tasks
-        # new dir with randomized name to prevent collision
-        dest_dir = "transfer_dest_dir-" + str(getrandbits(128))
-        dest_path = "/" + dest_dir + "/"
-        self.tc2.operation_mkdir(self.managed_ep_id, dest_path)
-
-        # transfer a file to the new dir
-        tdata = globus_sdk.TransferData(self.tc2, GO_EP1_ID,
-                                        self.managed_ep_id,
-                                        notify_on_succeeded=False)
-        source_path = "/share/godata/"
-        file_name = "file1.txt"
-        tdata.add_item(source_path + file_name, dest_path + file_name)
-        transfer_id = self.tc2.submit_transfer(tdata)["task_id"]
-
-        # delete the new dir
-        ddata = globus_sdk.DeleteData(self.tc2, self.managed_ep_id,
-                                      recursive=True,
-                                      notify_on_succeeded=False)
-        ddata.add_item(dest_path)
-        delete_id = self.tc2.submit_delete(ddata)["task_id"]
-
-        # sdktester1a gets endpoint manager task list
-        tasks_doc = self.tc.endpoint_manager_task_list(
-            filter_endpoint=GO_EP2_ID,
-            filter_user_id=get_user_data()["sdktester2b"]["id"])
-
-        # confirm submitted tasks can be found
-        # and tasks have some expected fields
-        expected_fields = ["username", "deadline", "type",
-                           "source_endpoint_id"]
-        delete_found = False
-        transfer_found = False
-        self.assertIsInstance(tasks_doc, PaginatedResource)
-        for task in tasks_doc:
-
-            for field in expected_fields:
-                self.assertIn(field, task)
-
-            if task["task_id"] == transfer_id:
-                transfer_found = True
-            if task["task_id"] == delete_id:
-                delete_found = True
-            if transfer_found and delete_found:
-                break
-
-        # fail if both not found
-        self.assertTrue(delete_found and transfer_found)
-
-    def test_endpoint_manager_get_task(self):
-        """
-        Has sdktester2b submit a no-op task on the managed endpoint
-        Confirms sdktester1a can view the task as an admin.
-        Confirms 403 when non manager attempts to use this resource.
-        """
-        # sdktester2b subits no-op delete task
-        ddata = globus_sdk.DeleteData(self.tc2, self.managed_ep_id,
-                                      notify_on_fail=False)
-        ddata.add_item("no-op.txt")
-        task_id = self.tc2.submit_delete(ddata)["task_id"]
-
-        # sdktester1a gets the task as admin
-        task_doc = self.tc.endpoint_manager_get_task(task_id)
-
-        self.assertEqual(task_doc["task_id"], task_id)
-        self.assertEqual(task_doc["owner_id"],
-                         get_user_data()["sdktester2b"]["id"])
-        self.assertEqual(task_doc["type"], "DELETE")
-        self.assertIn("status", task_doc)
-
-        # 403 for non managers, even if they submitted the task
-        with self.assertRaises(TransferAPIError) as apiErr:
-            self.tc2.endpoint_manager_get_task(task_id)
-        self.assertEqual(apiErr.exception.http_status, 403)
-        self.assertEqual(apiErr.exception.code, "PermissionDenied")
-
-    def test_endpoint_manager_task_event_list(self):
-        """
-        Has sdktester2b submit a no-op task on the managed endpoint.
-        Waits for task to fail, and confirms sdktester1a can see the
-        failure event as an admin.
-        Confirms 403 when non manager attempts to use this resource.
-        """
-        # sdktester2b subits no-op delete task and waits for completion
-        ddata = globus_sdk.DeleteData(self.tc2, self.managed_ep_id,
-                                      notify_on_fail=False)
-        ddata.add_item("no-op.txt")
-        task_id = self.tc2.submit_delete(ddata)["task_id"]
-        self.assertTrue(
-            self.tc2.task_wait(task_id, timeout=30, polling_interval=1))
-
-        # sdktester1a gets the task event list as admin
-        events_doc = self.tc.endpoint_manager_task_event_list(task_id)
-        self.assertIsInstance(events_doc, PaginatedResource)
-
-        failure_event = events_doc[0]  # most recent event is first
-        self.assertEqual(failure_event["DATA_TYPE"], "event")
-        self.assertEqual(failure_event["code"], "FILE_NOT_FOUND")
-        self.assertEqual(failure_event["description"],
-                         "No such file or directory")
-
-        # 403 for non managers, even if they submitted the task
-        with self.assertRaises(TransferAPIError) as apiErr:
-            self.tc2.endpoint_manager_task_event_list(task_id)
-        self.assertEqual(apiErr.exception.http_status, 403)
-        self.assertEqual(apiErr.exception.code, "PermissionDenied")
-
-    def test_endpoint_manager_task_successful_transfers(self):
-        """
-        Has sdktester2b submit a recursive transfer of share/godata to the
-        managed_ep. Waits for the task to complete, then has sdktester1a get
-        the successful transfers of the task as an admin. Confirms all 3 files
-        are seen, and some expected fields are present.
-        """
-        # new dir with randomized name to prevent collision
-        dest_dir = "transfer_dest_dir-" + str(getrandbits(128))
-        dest_path = "/" + dest_dir + "/"
-        self.tc2.operation_mkdir(self.managed_ep_id, dest_path)
-        # transfer a the files
-        tdata = globus_sdk.TransferData(self.tc2, GO_EP1_ID,
-                                        self.managed_ep_id,
-                                        notify_on_succeeded=False)
-        source_path = "/share/godata/"
-        tdata.add_item(source_path, dest_path, recursive=True)
-        task_id = self.tc2.submit_transfer(tdata)["task_id"]
-
-        # track asset for cleanup
-        self.asset_cleanup.append({"function": self.deleteHelper,
-                                   "args": [self.managed_ep_id, dest_path]})
-        # wait for task to complete
-        self.assertTrue(
-            self.tc2.task_wait(task_id, timeout=30, polling_interval=1))
-
-        # sdktester1a gets successful transfers as admin
-        success_doc = self.tc.endpoint_manager_task_successful_transfers(
-            task_id)
-
-        # confirm results
-        self.assertIsInstance(success_doc, PaginatedResource)
-        count = 0
-        for transfer in success_doc:
-
-            self.assertEqual(transfer["DATA_TYPE"], "successful_transfer")
-            self.assertIsNotNone(re.match(dest_path + "file[1-3].txt",
-                                          transfer["destination_path"]))
-            count += 1
-        self.assertEqual(count, 3)
-
-    def _unauthorized_transfers(self):
-        """
-        Helper that has sdktester2b submit 3 unauthorized transfers from the
-        managed endpoint, returns a list of their task_ids,
-        and tracks them for cleanup.
-        """
-        # submit the tasks
-        task_ids = []
-        for i in range(3):
-            tdata = globus_sdk.TransferData(self.tc2, self.managed_ep_id,
-                                            GO_EP1_ID, notify_on_fail=False)
-            tdata.add_item("/", "/", recursive=True)
-            task_ids.append(self.tc2.submit_transfer(tdata)["task_id"])
-
-        # track assets for cleanup
-        self.asset_cleanup.append(
-            {"function": self.tc.endpoint_manager_cancel_tasks,
-             "args": [task_ids, "Cleanup for unauthorized_transfers helper"]})
-
-        return task_ids
-
-    def test_endpoint_manager_cancel_tasks(self):
-        """
-        Get task ids from _unauthorized transfers, and has sdktester1a cancel
-        those tasks. Validates results.
-        Confirms 403 when non manager attempts to use this resource.
-        """
-        # cancel the tasks
-        task_ids = self._unauthorized_transfers()
-        message = "SDK test cancel tasks"
-        cancel_doc = self.tc.endpoint_manager_cancel_tasks(task_ids, message)
-
-        # validate results
-        self.assertEqual(cancel_doc["DATA_TYPE"], "admin_cancel")
-        self.assertIn("done", cancel_doc)
-        self.assertIn("id", cancel_doc)
-
-        # 403 for non managers, even if they submitted the tasks
-        with self.assertRaises(TransferAPIError) as apiErr:
-            self.tc2.endpoint_manager_cancel_tasks(task_ids, message)
-        self.assertEqual(apiErr.exception.http_status, 403)
-        self.assertEqual(apiErr.exception.code, "PermissionDenied")
-
-    def test_endpoint_manager_cancel_status(self):
-        """
-        Has sdktester2b submit three unauthorized transfers from the managed
-        endpoint, and sdktester1a admin_cancel those tasks.
-        Gets the cancel status of the cancel and validates results.
-        Loops while status is not done, then confirms all tasks canceled.
-        """
-        # cancel the tasks and get the cancel id
-        task_ids = self._unauthorized_transfers()
-        message = "SDK test cancel status"
-        cancel_id = self.tc.endpoint_manager_cancel_tasks(
-            task_ids, message)["id"]
-
-        # loop while not done or fail after 30 tries, 1 try per second
-        for tries in range(30):
-            # get and validate cancel status
-            status_doc = self.tc.endpoint_manager_cancel_status(cancel_id)
-            self.assertEqual(status_doc["DATA_TYPE"], "admin_cancel")
-            self.assertEqual(status_doc["id"], cancel_id)
-            if status_doc["done"]:
-                break
-            else:
-                time.sleep(1)
-
-        # confirm sdktester2b now sees all tasks as canceled by admin.
-        for task_id in task_ids:
-            task_doc = self.tc2.get_task(task_id)
-            self.assertEqual(task_doc["canceled_by_admin"], "SOURCE")
-            self.assertEqual(task_doc["canceled_by_admin_message"], message)
-
-    # TODO: stop skipping these tests when
-    # https://github.com/globusonline/koa/issues/49
-    # is resolved.
-    @unittest.skipIf(True, "github.com/globusonline/koa/issues/49")
-    def test_endpoint_manager_pause_tasks(self):
-        """
-        Has sdktester2b submit three unauthorized transfers,
-        and sdktester1a pause the tasks as an admin.
-        Validates results and confirms the tasks are paused.
-        Confirms 403 when non manager attempts to use this resource.
-        """
-        # pause the tasks
-        task_ids = self._unauthorized_transfers()
-        message = "SDK test pause tasks"
-        pause_doc = self.tc.endpoint_manager_pause_tasks(task_ids, message)
-
-        # validate results
-        self.assertEqual(pause_doc["DATA_TYPE"], "result")
-        self.assertEqual(pause_doc["code"], "PauseAccepted")
-
-        # confirm sdktester2b sees the tasks as paused
-        for task_id in task_ids:
-            task_doc = self.tc2.get_task(task_id)
-            self.assertTrue(task_doc["is_paused"])
-
-        # 403 for non managers
-        with self.assertRaises(TransferAPIError) as apiErr:
-            self.tc2.endpoint_manager_pause_tasks(task_ids, message)
-        self.assertEqual(apiErr.exception.http_status, 403)
-        self.assertEqual(apiErr.exception.code, "PermissionDenied")
-
-    @unittest.skipIf(True, "github.com/globusonline/koa/issues/49")
-    def test_endpoint_manager_resume_tasks(self):
-        """
-        Has sdktester2b submit three unauthorized transfers,
-        then sdktester1a pauses then resumes the tasks as an admin.
-        Confirms tasks go from paused to active.
-        Confirms 403 when non manager attempts to use this resource.
-        """
-        # pause the tasks and confirm they are paused
-        task_ids = self._unauthorized_transfers()
-        message = "SDK test resume tasks"
-        self.tc.endpoint_manager_pause_tasks(task_ids, message)
-        for task_id in task_ids:
-            task_doc = self.tc2.get_task(task_id)
-            self.assertTrue(task_doc["is_paused"])
-
-        # resume the tasks and validate results
-        resume_doc = self.tc.endpoint_manager_resume_tasks(task_ids, message)
-        self.assertEqual(resume_doc["DATA_TYPE"], "result")
-        self.assertEqual(resume_doc["code"], "ResumeAccepted")
-
-        # confirm tasks are now active.
-        for task_id in task_ids:
-            task_doc = self.tc2.get_task(task_id)
-            self.assertFalse(task_doc["is_paused"])
-
-        # 403 for non managers
-        with self.assertRaises(TransferAPIError) as apiErr:
-            self.tc2.endpoint_manager_resume_tasks(task_ids, message)
-        self.assertEqual(apiErr.exception.http_status, 403)
-        self.assertEqual(apiErr.exception.code, "PermissionDenied")
-
-    def test_add_endpoint_role(self):
-        """
-        Adds a role to the test share endpoint, validates results
-        returns role_id for use in get and delete
-        """
-        # add the new role
-        add_data = {"principal_type": "identity",
-                    "principal": get_user_data()["go"]["id"],
-                    "role": "access_manager"
-                    }
-        add_doc = self.tc.add_endpoint_role(self.managed_ep_id, add_data)
-        role_id = add_doc["id"]
-
-        # track asset for cleanup
-        self.asset_cleanup.append({"function": self.tc.delete_endpoint_role,
-                                   "args": [self.managed_ep_id, role_id],
-                                   "name": "test_role"})
-
-        # validate results
-        for key in add_data:
-            self.assertEqual(add_doc[key], add_data[key])
-
-        # return role id
-        return role_id
-
-    def test_get_endpoint_role(self):
-        """
-        Gets role created in test_add_endpoint_role, validates results.
-        """
-        role_id = self.test_add_endpoint_role()
-        get_doc = self.tc.get_endpoint_role(self.managed_ep_id, role_id)
-
-        # validate results
-        self.assertEqual(get_doc["DATA_TYPE"], "role")
-        self.assertEqual(get_doc["id"], role_id)
-        self.assertEqual(get_doc["principal_type"], "identity")
-        self.assertEqual(get_doc["principal"],
-                         get_user_data()["go"]["id"])
-        self.assertEqual(get_doc["role"], "access_manager")
-
-    def test_delete_endpoint_role(self):
-        """
-        Deletes role created in test_add_endpoint_role, validates results.
-        """
-        role_id = self.test_add_endpoint_role()
-        delete_doc = self.tc.delete_endpoint_role(self.managed_ep_id, role_id)
-
-        # validate results
-        self.assertEqual(delete_doc["DATA_TYPE"], "result")
-        self.assertEqual(delete_doc["code"], "Deleted")
-        self.assertIn("deleted successfully", delete_doc["message"])
-
-        # confirm get no longer sees the role
-        with self.assertRaises(TransferAPIError) as apiErr:
-            self.tc.get_endpoint_role(self.managed_ep_id, role_id)
-        self.assertEqual(apiErr.exception.http_status, 404)
-        self.assertEqual(apiErr.exception.code, "RoleNotFound")
-
-        # stop tracking asset for cleanup
-        for cleanup in self.asset_cleanup:
-            if "name" in cleanup and cleanup["name"] == "test_role":
-                self.asset_cleanup.remove(cleanup)
-                break
